Estudo_next/testes/backend/requirements-dev.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.1
django-cors-headers==3.13.0
django-discover-runner==1.0
django-extensions==3.2.0
django-ninja==0.19.1
django-role-permissions==3.1.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
idna==3.3
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.14.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
oauthlib==3.2.0
orjson==3.8.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.2
Pygments==2.13.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-pkcs12==1.14
Send2Trash==1.8.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.0
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.2
